==> Dollar Tree/2. Joining_dfs_Violation_Summary.ipynb <==
import pandas as pd
pd.options.display.max_colwidth = None
pd.options.display.max_columns = None

df1 = pd.read_csv("1_activities_df.csv")

df2 = pd.read_csv("1_details_df.csv")

df3 = pd.read_csv("1_violation_summary_df.csv")

df3

master_df = pd.merge(df1, df2, how='inner', left_on='Activity', right_on='Inspection Nr')

master_df

final_df = pd.merge(master_df, df3, how='inner', left_on='Activity', right_on='activity_code')

final_df

final_df.to_csv('2_final_df.csv')

!ls
