==> Dollar Tree/5. Work2_MoreWork_Violations.ipynb <==
import pandas as pd

pd.options.display.max_colwidth = None
pd.options.display.max_columns = None

!ls

### 1. For Current Violations

df = pd.read_csv("3_currentviolationsdrop.csv")

df

df['year'] = pd.to_datetime(df['Open Date'], format='%m/%d/%Y')

df['simple_year'] = df['year'].dt.year

df.head(20)

df.dtypes

df['Total'] = df['Total'].astype(int)

violations_df = df.groupby('simple_year')['Total'].sum()

violations_df

new2_df = violations_df.to_frame(name='violations')

new2_df

new2_df.to_csv('5_CurrentViolations.csv')

### 2. For Initial Violations

df = pd.read_csv("3_initialviolationsdrop.csv")

df['year'] = pd.to_datetime(df['Open Date'], format='%m/%d/%Y')

df['simple_year'] = df['year'].dt.year

df['Total'] = df['Total'].astype(int)

violations_df = df.groupby('simple_year')['Total'].sum()

violations_df

new2_df = violations_df.to_frame(name='violations')

new2_df

new2_df.to_csv('5_InitialViolations.csv')
